requirements.txt
streamlit
pandas
plotly
pytz
supabase
python-dotenv
